+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- implement Prometheus exporter that provides all quota results
+
+### Changed
+
+- display AWS account ID instead of profile name in check scope
+
 ## [1.1.0] - 2021-02-27
 
 ### Added